// src/api/generated/models/consolidationSummary.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { ConsolidationSummaryStatus } from './consolidationSummaryStatus';
import type { ConsolidationSummaryBuyerCompany } from './consolidationSummaryBuyerCompany';
import type { ConsolidationSummaryTransportMode } from './consolidationSummaryTransportMode';
import type { ConsolidationSummaryCarrier } from './consolidationSummaryCarrier';
import type { ConsolidationSummaryTrackingNumber } from './consolidationSummaryTrackingNumber';
import type { ConsolidationSummaryOrigin } from './consolidationSummaryOrigin';
import type { ConsolidationSummaryDestination } from './consolidationSummaryDestination';
import type { ConsolidationSummaryPlannedAt } from './consolidationSummaryPlannedAt';
import type { ConsolidationSummaryInTransitAt } from './consolidationSummaryInTransitAt';
import type { ConsolidationSummaryCustomsAt } from './consolidationSummaryCustomsAt';
import type { ConsolidationSummaryDeliveredAt } from './consolidationSummaryDeliveredAt';

export interface ConsolidationSummary {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  status: ConsolidationSummaryStatus;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  buyerCompanyId: string;
  buyerCompany: ConsolidationSummaryBuyerCompany;
  transportMode: ConsolidationSummaryTransportMode;
  carrier: ConsolidationSummaryCarrier;
  trackingNumber: ConsolidationSummaryTrackingNumber;
  origin: ConsolidationSummaryOrigin;
  destination: ConsolidationSummaryDestination;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  shippingInvoiceCount: number;
  plannedAt: ConsolidationSummaryPlannedAt;
  inTransitAt: ConsolidationSummaryInTransitAt;
  customsAt: ConsolidationSummaryCustomsAt;
  deliveredAt: ConsolidationSummaryDeliveredAt;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  createdAt: string;
}