// src/api/generated/models/consolidationSummaryStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type ConsolidationSummaryStatus = typeof ConsolidationSummaryStatus[keyof typeof ConsolidationSummaryStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ConsolidationSummaryStatus = {
  DRAFT: 'DRAFT',
  PLANNED: 'PLANNED',
  IN_TRANSIT: 'IN_TRANSIT',
  CUSTOMS: 'CUSTOMS',
  DELIVERED: 'DELIVERED',
} as const;